[iOS8Upgrades] Refactored the whole solution in order to provide a cleaner, much more transparent approach to solve the same issue at hand: facilitate cell height based on whether text view sizes and provide scrolling while editing a text view. The refactor takes into account the fact that iOS 8 introduced new ways to calculate cell heights based on auto layout. We will take this to our advantage now. Solution is working in the example app, but requires some refurbishments. Fixed all auto layout warnings by applying a constraint priority of 800.
Missing: Scrolling while editing is not working in some cases, tableView updates (beginUpdates / endUpdates) are making the tableView reload so it's scrolling to the top when i.e pressing return (carriage return on a text view). Additional to the latter key functionalities, some improvements need to be made: simplify where necessary, clean code, use protocols in example UITableViewCell subclass, provide a way to avoid the manager to be an intermediary for TextViewDelegate callbacks, and instead allow the client to manually update if necessary. Amongst other things, seems good so far!